+++ extensions/ql-vscode/src/vscode-tests/factories/remote-queries/shared/variant-analysis-submission.ts
@@ -1,0 +1,20 @@
+import { VariantAnalysisQueryLanguage, VariantAnalysisSubmission } from '../../../../remote-queries/shared/variant-analysis';
+
+export function createMockSubmission(): VariantAnalysisSubmission {
+  return {
+    startTime: 1234,
+    controllerRepoId: 5678,
+    actionRepoRef: 'repo-ref',
+    query: {
+      name: 'query-name',
+      filePath: 'query-file-path',
+      language: VariantAnalysisQueryLanguage.Javascript,
+      pack: 'base64-encoded-string',
+    },
+    databases: {
+      repositories: ['1', '2', '3'],
+      repositoryLists: ['top10', 'top100'],
+      repositoryOwners: ['mona', 'lisa'],
+    }
+  };
+}
